+++ src/compas/geometry/vector.py
@@ -148,11 +148,25 @@ class Vector(Geometry):
     def __sub__(self, other):
         return Vector(self.x - other[0], self.y - other[1], self.z - other[2])
 
-    def __mul__(self, n):
-        return Vector(self.x * n, self.y * n, self.z * n)
+    def __mul__(self, other):
+        if isinstance(other, (int, float)):
+            return Vector(self.x * other, self.y * other, self.z * other)
 
-    def __truediv__(self, n):
-        return Vector(self.x / n, self.y / n, self.z / n)
+        try:
+            other = Vector(*other)
+            return Vector(self.x * other.x, self.y * other.y, self.z * other.z)
+        except TypeError:
+            raise TypeError("Cannot cast {} {} to Vector".format(other, type(other)))
+
+    def __truediv__(self, other):
+        if isinstance(other, (int, float)):
+            return Vector(self.x / other, self.y / other, self.z / other)
+
+        try:
+            other = Vector(*other)
+            return Vector(self.x / other.x, self.y / other.y, self.z / other.z)
+        except TypeError:
+            raise TypeError("Cannot cast {} {} to Vector".format(other, type(other)))
 
     def __pow__(self, n):
         return Vector(self.x**n, self.y**n, self.z**n)
@@ -189,6 +203,26 @@ class Vector(Geometry):
         self.y **= n
         self.z **= n
         return self
+
+    def __rmul__(self, n):
+        return self.__mul__(n)
+
+    def __radd__(self, other):
+        return self.__add__(other)
+
+    def __rsub__(self, other):
+        try:
+            other = Vector(*other)
+            return other - self
+        except TypeError:
+            raise TypeError("Cannot cast {} {} to Vector".format(other, type(other)))
+
+    def __rtruediv__(self, other):
+        try:
+            other = Vector(*other)
+            return other / self
+        except TypeError:
+            raise TypeError("Cannot cast {} {} to Vector".format(other, type(other)))
 
     # ==========================================================================
     # Properties